Update to Namespace and Directorys
+ To fix potential conflicts with original library
